added test server
